File: SRC/TRF3dV100/dsparseTreeFactorGPU_impl.hpp Refactor dsparseTreeFactorGPU_impl.hpp
Files Changed:
- dsparseTreeFactorGPU_impl.hpp

Detailed Changes:
- Added the `<cinttypes>` header file.
- Commented out the `printf` statement and replaced it with a new `printf` statement that uses the `PRIdMAX` format specifier to print the value of `winSize`.

File: SRC/TRF3dV100/lupanels_impl.hpp
Refactor `lupanels_impl.hpp` file in SRC/TRF3dV100 directory

Files Changed:
- `lupanels_impl.hpp`: Refactored the `initDiagFactBufsArr` and `freeDiagFactBufsArr` functions to use `SUPERLU_MALLOC` instead of `new` for memory allocation. Also, updated the `getBigV` function to use `(Ftype*) SUPERLU_MALLOC` for memory allocation.

Detailed Changes:
- Updated the `initDiagFactBufsArr` function to use `SUPERLU_MALLOC` for allocating memory for `dFBufs` array and `BlockUFactor` and `BlockLFactor` variables.
- Updated the `freeDiagFactBufsArr` function to free memory using `SUPERLU_FREE` for `BlockUFactor`, `BlockLFactor`, and `dFBufs` variables.
- Updated the `getBigV` function to use `(Ftype*) SUPERLU_MALLOC` for allocating memory for `bigV` variable.

File: SRC/TRF3dV100/xlupanels.hpp
Refactor xLUstruct_t in xlupanels.hpp

Files Changed:
- xlupanels.hpp

Detailed Changes:
- Commented out the call to dfreeDiagFactBufsArr() and added a new call to freeDiagFactBufsArr() in xLUstruct_t struct.
- Added a new function freeDiagFactBufsArr() to free diagonal factorization buffers.
- Added two new functions copyLpanelsToGPU() and copyUpanelsToGPU() to copy L and U panels to the GPU.
- Added a new function dDFactPSolveGPU() to perform diagonal factorization on the GPU.

+++ SRC/TRF3dV100/dsparseTreeFactorGPU_impl.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include <cstdio>
+#include <cinttypes>
 #include "superlu_ddefs.h"
 #include "lupanels.hpp"
 #ifdef HAVE_CUDA
@@ -475,7 +476,8 @@ int_t xLUstruct_t<Ftype>::dsparseTreeFactorGPU(
     //  everything else can remain as it is
     int_t winSize = SUPERLU_MIN(numLA / 2, eTreeTopLims[1]);
 
-    printf(". lookahead winSize %d\n", winSize);
+    // printf(". lookahead winSize %d\n", winSize);
+    printf(". lookahead winSize %" PRIdMAX "\n", static_cast<intmax_t>(winSize));
     fflush(stdout);
 
     for (int k0 = k_st; k0 < winSize; ++k0)
